requirements.txt
asgiref==3.7.2
Django==5.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
PyJWT==2.5.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
